9.Modules/Exercise/requirements.txt
frame==0.3.1
Jinja2==3.1.2
MarkupSafe==2.1.2
multipart==0.2.4
Pillow==9.5.0
pyfiglet==0.8.post1
pytz==2023.3
repoze.lru==0.7
Routes==2.5.1
six==1.16.0
